feat: add comparison toggle button
