+++ src/App.jsx
@@ -1,13 +1,13 @@
-import React from 'react';
+import React from "react";
 import {
   BrowserRouter as Router,
   Route,
   Routes,
   Navigate,
-} from 'react-router-dom';
+} from "react-router-dom";
 
-import LoginView from './views/loginView/loginView';
-import MainView from './views/mainView/mainView';
+import LoginView from "./views/loginView/loginView";
+import MainView from "./views/mainView/mainView";
 
 function App() {
   return (
