+++ src/components/charts/dongCharts/salesTime.jsx
@@ -24,45 +24,49 @@ const SalesTime = ({ code }) => {
           chart.destroy(); // 이전 차트가 있으면 파괴
         }
 
-        const ctx = chartRef.current.getContext("2d");
-        const newChart = new Chart(ctx, {
-          
-          data: {
-            labels: labels,
-            datasets: [
-              {
-                type: "line",
-                data: values,
-                backgroundColor: "rgba(75, 192, 192, 0.5)",
-                borderColor: "rgb(75, 192, 192)",
-                borderWidth: 1,
-              }, {
-                type: "bar",
-                data: values
-              }
-            ],
-          },
-          options: {
-            scales: {},
-            plugins: {
-              legend: {
-                display: false, // 범례 표시
-              },
-              title: {
-                display: true,
-                text: "행정동 시간별 매출", // 차트 타이틀
+        if (chartRef.current) {
+          const ctx = chartRef.current.getContext("2d");
+          const newChart = new Chart(ctx, {
+            data: {
+              labels: labels,
+              datasets: [
+                {
+                  type: "line",
+                  data: values,
+                  backgroundColor: "rgba(75, 192, 192, 0.5)",
+                  borderColor: "rgb(75, 192, 192)",
+                  borderWidth: 1,
+                },
+                {
+                  type: "bar",
+                  data: values,
+                },
+              ],
+            },
+            options: {
+              scales: {},
+              plugins: {
+                legend: {
+                  display: false, // 범례 표시
+                },
+                title: {
+                  display: true,
+                  text: "행정동 시간별 매출", // 차트 타이틀
+                },
               },
             },
-          },
-        });
+          });
 
-        setChart(newChart);
+          setChart(newChart);
+        }
       } catch (error) {
         console.error("Failed to fetch area data", error);
       }
     };
 
-    fetchData();
+    if (chartRef.current) {
+      fetchData();
+    }
   }, [code]); // code가 변경될 때마다 fetchData 실행
 
   return (
