+++ src/components/charts/dongComparisonCharts/industrySalesComparison.jsx
@@ -7,26 +7,52 @@ const IndustrySalesComparison = ({ code1, code2, category }) => {
   const [chart, setChart] = useState(null);
   const [isEmpty, setIsEmpty] = useState(false);
 
+  //같은 업종 매출 비교
   useEffect(() => {
     const fetchData = async () => {
       try {
         const [response1, response2] = await Promise.all([
-          axios.get(`https://api.gadduck.info/towns/industry/sales?code=${code1}&name=${category}`),
-          axios.get(`https://api.gadduck.info/towns/industry/sales?code=${code2}&name=${category}`)
+          axios.get(
+            `https://api.gadduck.info/towns/industry/sales?code=${code1}&name=${category}`
+          ),
+          axios.get(
+            `https://api.gadduck.info/towns/industry/sales?code=${code2}&name=${category}`
+          ),
         ]);
 
         const salesData1 = response1.data.data;
         const salesData2 = response2.data.data;
 
-        if (!salesData1 || Object.keys(salesData1).length === 0 || !salesData2 || Object.keys(salesData2).length === 0) {
-          setIsEmpty(true); 
+        if (
+          !salesData1 ||
+          Object.keys(salesData1).length === 0 ||
+          !salesData2 ||
+          Object.keys(salesData2).length === 0
+        ) {
+          setIsEmpty(true);
           return;
         }
 
         setIsEmpty(false);
 
-        const originalLabels = ["월요일", "화요일", "목요일", "금요일", "토요일", "일요일", "수요일"];
-        const labels = ["월요일", "화요일", "수요일", "목요일", "금요일", "토요일", "일요일"];
+        const originalLabels = [
+          "월요일",
+          "화요일",
+          "수요일",
+          "목요일",
+          "금요일",
+          "토요일",
+          "일요일",
+        ];
+        const labels = [
+          "월요일",
+          "화요일",
+          "수요일",
+          "목요일",
+          "금요일",
+          "토요일",
+          "일요일",
+        ];
 
         const reorderData = (data) => {
           const reordered = [];
@@ -37,8 +63,12 @@ const IndustrySalesComparison = ({ code1, code2, category }) => {
           return reordered;
         };
 
-        const counts1 = reorderData(Object.values(salesData1).map(count => parseFloat(count)));
-        const counts2 = reorderData(Object.values(salesData2).map(count => parseFloat(count)));
+        const counts1 = reorderData(
+          Object.values(salesData1).map((count) => parseFloat(count))
+        );
+        const counts2 = reorderData(
+          Object.values(salesData2).map((count) => parseFloat(count))
+        );
 
         if (chart) {
           chart.destroy();
@@ -74,7 +104,7 @@ const IndustrySalesComparison = ({ code1, code2, category }) => {
                 text: `행정동 별 ${category} 매출 비교`,
               },
               legend: {
-                position: 'top',
+                position: "top",
               },
             },
             scales: {
