+++ src/components/labels/industryBusiness.jsx
